# docs/swagger/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  jpcorrect-backend_internal_domain.ReportThemeSuggestion:
    properties:
      created_at:
        type: string
      description:
        type: string
      suggestion_id:
        type: string
      title:
        type: string
    type: object
  jpcorrect-backend_internal_domain.Topic:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      difficulty:
        type: string
      hint_grammar:
        items:
          type: integer
        type: array
      hint_vocab:
        items:
          type: integer
        type: array
      kind:
        $ref: '#/definitions/jpcorrect-backend_internal_domain.TopicKind'
      title_jp:
        type: string
      topic_id:
        type: string
      updated_at:
        type: string
    type: object
  jpcorrect-backend_internal_domain.TopicKind:
    enum:
    - announce
    - random
    type: string
    x-enum-varnames:
    - TopicKindAnnounce
    - TopicKindRandom
...
host: localhost:8080
info:
  contact: {}
  description: Japanese language correction platform backend API
  title: jpcorrect API
  version: "1.0"
paths:
...
  /v1/topics/{id}:
    get:
      description: Get a single topic with hint_vocab and hint_grammar
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jpcorrect-backend_internal_domain.Topic'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a topic by ID
      tags:
      - topics
  /v1/topics/announce/search:
    get:
      description: Search announce topics by keyword
      parameters:
      - description: Search keyword
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/jpcorrect-backend_internal_domain.Topic'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Search announce topics
      tags:
      - topics
  /v1/topics/random:
    get:
      description: Get a random topic
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jpcorrect-backend_internal_domain.Topic'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a random topic
      tags:
      - topics
  /v1/topics/report-suggestions:
    get:
      description: Get all report theme suggestions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/jpcorrect-backend_internal_domain.ReportThemeSuggestion'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List report theme suggestions
      tags:
      - topics
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"